src/vertex_training/requirements.txt
lightgbm>=3.3.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
joblib>=1.2.0
